requirements.txt
google-cloud-storage==2.5.0